+++ NEWS.md
@@ -5,8 +5,10 @@
 * Updated 'color' throughout (#11).
 * Added `cr_choose_bw()` to README.
 * Added code of conduct.
+* Add basic stopifnots.
 
 # coloratio 0.0.0.9000
 
 * Added `get_ratio()` function.
 * Added NEWS, LICENSE, README.
+1